requirements.txt
streamlit
chromadb
sentence-transformers
pdfplumber
python-dotenv
requests
pymongo
json-repair

langchain
langchain-community
langchain-openai
langchain-core
langchain-huggingface

openai
torchvision
pypdf
transformers
torch
reportlab
nltk
